Fix circular imports for ProfileManagement
